slightly change firebase file structure
